refactor: add type file to toolbox
